+++ frontend/src/App.js
@@ -19,8 +19,8 @@ import ProductListScreen from "./screens/AdminScreens/ProductListScreen";
 import ProductEditScreen from "./screens/AdminScreens/ProductEditScreen";
 import OrderListScreen from "./screens/AdminScreens/OrderListScreen";
 import ThankYouScreen from "./screens/Payment/ThankYouScreen";
-import HomeScreen from "./screens/HomeScreen";
-import AvionicsScreen from "./screens/AvionicsScreen";
+import HomeScreen from "./components/Home/HomeScreen";
+import AvionicsScreen from "./components/Avionics/AvionicsScreen";
 import MaintenanceScreen from "./screens/MaintenanceScreen";
 import FlirScreen from "./screens/FlirScreen";
 import TeamScreen from "./screens/TeamScreen";
